api_gateway/requirements.txt
fastapi
grpcio
grpcio-tools